Materialize functions with forward-referenced blocks in streamed modules
This fixes the assertion in the GlobalMerge pass which was
seeing GVs representing the unresolved forward references.
materializeForwardReferencedFunctions materializes these functions and removes
the special GVs from the global list.
If there is a function with forward referenced blocks near the end of the
bitcode file, then streaming could be ineffective for that file.

This CL should also be committed upstream, but will go in here so we can fix the
build before the next merge.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2882
TEST=computed_goto_test

Review URL: http://chromiumcodereview.appspot.com/10700165/