Add @nacl.read.tp() intrinsic, a fast version of NaCl's tls_get() IRT interface
This is in preparation for adding an LLVM pass that will expand out
TLS (thread_local) variable accesses into calls to nacl.read.tp.

On ARM, there is already an arm.thread.pointer intrinsic.  We reuse
the code for that.

On x86, we have to add an implementation.  The added code is based on
x86's LowerToTLSExecModel() for the %gs:0 case, and on NaCl-MIPS'
LowerGlobalTLSAddress() for the __nacl_read_tp() case.  (In contrast,
X86NaClRewritePass.cpp inserts a __nacl_read_tp() call at the lower MI
level; we don't use that approach here.)

We convert LowerINTRINSIC_WO_CHAIN() into a method in order to access
the Subtarget member.  This is consistent with other x86 Lower methods
and with the ARM version.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2837
TEST="llvm-lit test/NaCl"

Review URL: https://codereview.chromium.org/11383002