Leave NOPS as upstream defines
Upstream tests fail because NOPS are expected to be 0xe1a00000 rather than
what we substitute.  This seems to have been a performance only fix, and the
ARM manuals are quite ambiguous about which to prefer for performance anyway.
I am reverting to the upstream version.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1711

Review URL: https://codereview.chromium.org/11280193