Limit the size of the bitcode download buffer on the untrusted side.
Previously if compilation was slower than download, bitcode waiting to be
compiled would accumulate in the SRPCStreamer buffer in llc. This CL limits
the size of that buffer and instead causes the RPC called from the browser
to block until there is room (i.e., the compilation has consumed enough
to cach up). This reduces the memory requirements on the untrusted side, and
also makes it possible for the browser plugin to get an estimate of the
progress of the compilation

R=jvoung@chromium.org,robertm@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/11360087