Add an RPC to LLC.nexe to init with appended flags (vs set all flags).
This will take the default flags and append any extra flags.
For now, this is only intended to support -O in the browser,
and perhaps -mcpu=.*, -mattr=.  We move the tuning parameters
like -mcpu out of the default, so that a user does not end up
setting the -mcpu more than once (which is disallowed).

What does go into the default is -mtriple, which we may want
to hard-code into the build somehow (but it's not yet clear
how to do that cleanly).  That way, a user does not need
to specify that portion.  We also hardcode the -sfi-* flags.

Sort of tested by: https://codereview.chromium.org/12459004/

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3325

Review URL: https://codereview.chromium.org/12490004