Add -time-ir-parsing flag to llc.
This is in the process of being added upstream but the exact location
is still debated and will take some time to settle. In the mean-time
I'm adding it as a localmod so we can run our benchmarks effectively.
Later this should be switched to the upstream implementation (or move
to our own llc-replacement-driver).

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3349