PNaCl: Allow the ABI checker to be used from "opt"
This allows the ABI checker passes to be used in the same way as
LLVM's "-verify" pass.  It allows the checker to be run between other
passes, and without launching pnacl-abicheck as a separate process (so
without the overhead of reading bitcode into memory again).

Make the ABI checker passes produce fatal errors by default, to match
"-verify".  This is overridden for pnacl-abicheck's use.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2309
TEST=tested with pnacl-ld.py changes to use the ABI checker passes

Review URL: https://codereview.chromium.org/13323006