PNaCl: Add FlattenGlobals pass for simplifying global variable initializers
This pass converts initializers for global variables into a flattened
normal form which removes nested struct types and simplifies
ConstantExprs.

In the future, we might change the bitcode format to use the flattened
bytes+relocs representation for global initializers.  In that case, we
will be able to reuse the FlattenedConstant class in the bitcode
reader/writer for converting to and from this form.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3113
TEST=test/Transforms/NaCl/flatten-globals.ll

Review URL: https://codereview.chromium.org/14017011