Use unique'ed types for varargs expansion instead of non-unique named types.
Saves a tiny bit of space for var-args heavy bitcode
programs, since the anonymous types get unique'ed.  E.g.,
saves 20KB out of 1.6MB in spec gcc when comparing the
gzipped files, or about 100KB when not zipped.  This is only
a savings with the current wire format.  If we change the
alloca, etc. to only have sizes and not struct types
then we would also not have this duplication.

Just happened to notice while looking through code for
what struct types remain used in the bitcode.

random cleanup for:
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3338
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/14197004