Start to disallow llvm.frameaddress and llvm.returnaddress in ABI checker.
They do not seem to be widely used by user code.

* The boehm garbage collector library does reference
__builtin_return_address under an ifdef, but it
does not appear to be compiled in.

* Mesa-7.6 uses __builtin_frame_address
for u_debug_stack.c, but that also does not appear to be
part of the built libraries.

They expose stack/code addresses (at least the lower
32-bits of the address).

As part of https://codereview.chromium.org/14619022/,
we stopped considering the scons and gcc torture tests that
use these intrinsics as meeting the stable ABI.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3378
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/14657017