PNaCl: Add "-pnacl-abi-simplify-{pre,post}opt" meta-passes to "opt"
These meta-passes will be used to replace the pass lists that are
currently in the pnacl-ld.py driver script in the NaCl repo.

I've moved the comments across from pnacl-ld.py and added a couple
more comments for ExpandByVal and StripMetadata.

Fix the declaration of createResolveAliasesPass().

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3435
TEST=new *.ll tests + tested with change to pnacl-ld.py

Review URL: https://codereview.chromium.org/15669002