Move llvm.expect from dev to blacklist.  Convert with -lower-expect.
The backend currently treats llvm.expect as a nop
pass-through. Until it becomes more useful, we don't need
it in stable bitcode.  It sounds like the backend may
prefer to use the !prof metadata instead.

The -lower-expect pass will convert it into branch
weights, which can be consumed by the middle end optimizer's
block placement pass.

It is already converted to !prof metadata when clang
is run with -O2, but not when clang is run at -O0.

Also move the llvm.frameaddress from the dev part of
the intrinsics test to the disallowed part of the test.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3378
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/16882002