Rewrite llvm.flt.rounds to "1" for now, and disallow llvm.flt.rounds.
Until there is an intrinsic to *set* the rounding mode,
this intrinsic to *get* the rounding mode isn't so useful.
Separately we will add a test that for each platform, the
initial rounding mode is "1" (round to nearest).  That is
the case right now for x86, ARM, and MIPS.
(see https://codereview.chromium.org/16785003/)

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3491
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/17229007