PNaCl wire format: Remove the top-level TRIPLE record
Currently, the triple that gets put into pexes is
"armv7-none-linux-gnueabi".  This is a wart that we don't want to
keep.  We can remove the triple entirely; PNaCl doesn't need it.

We don't need to add a triple back at read time: (pnacl-)llc tells the
backend explicitly what target to use.

I'm leaving MODULE_CODE_TRIPLE defined in the enum for now because
pnacl-bcanalyzer still refers to it.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3505
TEST=PNaCl toolchain trybots + GCC torture tests + Spec2k

Review URL: https://codereview.chromium.org/17321009