Fix ARM LDMIA MI.
The instruction that was generated for paired register stack slot load was wrong. This was fixed by Tim Northover in LLVM 3.3 commit 179977, but his patch does much more and doesn't apply as-is to our tree. I'll therefore punt applying the full patch to when we rebase to 3.3.

I encountered the issue while working on atomics (64-bit atomics require paired registers on ARM), and saw the 3.3 fix when I tried upstreaming my fix.

BUG= non
TEST= ./pnacl/test.sh test-arm
R=eliben@chromium.org

Review URL: https://codereview.chromium.org/18699004