Remove the bitcast (of global values) from load instructions.
Adds the eliding of bitcasts that are used as an argument to
instructions that expect normalized pointers. Currently, the checked
in code only checks normalized pointers for load instructions. Hence,
the restriction to load instructions. As more instructions are
modified to check for normalized pointers, this code will apply
to those instructions.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3544
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/21614002