Updating the developer's guide with additional PNaCl information.
I don't think this specific format will be what we eventually "ship" to the NaCl
documentation website, but it's useful to have a placeholder to add things to.
For example, information like language and library versions supported,
preprocessor macros and so on. We will have to document these at some point.

Feel free to suggest additional such things that need to be documented (here or
in subsequent CLs).

BUG=None
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/23105006