PNaCl bitcode reader: Disallow pointer-typed arguments in indirect calls
Make the bitcode reader stricter, so that it disallows pointer
arguments in indirect function calls, which are disallowed by the
PNaCl ABI checker.  Pointer arguments in function calls are only
allowed in intrinsic calls, and calls to intrinsics must always be
direct calls, not indirect calls.

This involves removing two tests that specifically test for pointer
args.

This is in preparation for tweaking how indirect calls are
represented, so that they store the call's return type rather than the
function type.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3544
TEST=*.ll tests

Review URL: https://codereview.chromium.org/23660005