PNaCl bitcode: Don't output the LABEL type in the type table
The type ID for the "label" type is never referenced.

The "label" type was only being added to the type table because
EnumerateType() treats BasicBlock and Value operands the same.
However, the rest of the reader and writer treat BasicBlock operands
specially and not like other Values.

Change some tests to use wildcards for some type IDs.  This is so that
I don't have to update all the type ID numbers now that the generated
type tables have changed.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3590
TEST=run small_tests with v2 bitcode format enabled

Review URL: https://codereview.chromium.org/23530031