PNaCl bitcode: Indirect calls: Store return type instead of function type
For indirect call instructions (INST_CALL_INDIRECT), it's not
necessary to store the full function type.  The argument types are
already known from the arguments in the instruction.  We only need to
store the return type to be able to reconstruct the full function
type.

Storing only the return type ID will make the bitcode a little more
compact.  Return type IDs will be frequently-used scalar types, which
can be given smaller type IDs than function types, which are less
frequently used.

This potentially makes the writer simpler: In principle, the writer no
longer needs to make a pass across all functions' bodies to determine
which function types are used in order to build the type table.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3544
TEST=*.ll tests

Review URL: https://codereview.chromium.org/23521005