PNaCl bitcode: Strip pointer types from intrinsic declarations' parameters
Change the writer to strip pointer types from intrinsics' argument and
return types, replacing them with i32.  This simplifies the PNaCl
bitcode format so that pointer types don't need to be represented
here.

Change the reader to restore the pointer types so that the intrinsic
declarations pass the LLVM and PNaCl verifiers.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3671
TEST=intrinsic tests in call-elide.ll
  + intrinsic-pointer-args.ll
  + run small_tests with bitcode v2 enabled

Review URL: https://codereview.chromium.org/23793005