PNaCl bitcode: Change FORWARDTYPEREF to never use pointer types
Before, FORWARDTYPEREFs used i8* type rather than i32 if they
referenced an "alloca" instruction.

Clean this up so that FORWARDTYPEREFs use i32 instead in this case.

Note that this means that a forward-referenced "alloca" can be used
via an ptrtoint+inttoptr, rather than a bitcast, but that's no
problem.

This is a step towards removing TYPE_CODE_POINTER from the types
table, to simplify the PNaCl bitcode format.

Rename NormalizeParamType() to NormalizeScalarType() to reflect that
it's used in more cases; make it public.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3671
TEST=test/NaCl/Bitcode/*.ll

Review URL: https://codereview.chromium.org/23719016