PNaCl bitcode: Simplify how the writer elides casts
The PNaCl bitcode writer had some complex logic for deciding when to
omit casts in the output.  This was left over from when the writer was
trying to leave in the casts in some but not all cases (as part of
incrementally removing casts).

This is no longer needed now that the writer just omits all inttoptrs,
all ptrtoints, and all pointer bitcasts.

This cleanup also fixes the writer so that it elides an inttoptr of a
ptrtoint.  This sequence is allowed by the PNaCl ABI verifier, but
never occurred in practice because ReplacePtrsWithInts'
SimplifyCasts() function converts this sequence to an equivalent
bitcast.  Before this change, the writer would give this error for an
inttoptr-of-ptrtoint:

  LLVM ERROR: Illegal (PNaCl ABI) pointer cast :   %1 = ptrtoint [4 x i8]* @bytes to i32

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3590
TEST=toolchain trybots

Review URL: https://codereview.chromium.org/25817002