Apply upstream: [mips] Fix a bug in MipsLongBranch::replaceBranch
Cherry-pick r191978 from upstream.

Original commit message:

Author: Akira Hatanaka <ahatanaka@mips.com>
Date:   Fri Oct 4 20:51:40 2013 +0000

[mips] Fix a bug in MipsLongBranch::replaceBranch, which was erasing
instructions in delay slots along with the original branch instructions

This has to be cherrypicked, as it is a bug in backend. It was exposed in
a long function inside of llc, which caused llc.nexe to work incorrectly.

TBR= mseaborn@chromium.org, dschuff@chromium.org
BUG= bug in MIPS backend

Review URL: https://codereview.chromium.org/26933005