Fix bug in rewriting of library calls to intrinsics + new regression test.
The pass gets confused in some cases when library functions get passed to
other functions as arguments, because use_iterator returns the call instruction.
The existing test (rewrite-longjmp-noncall-uses.ll) did not catch this problem
because there a bitcast constexpr was applied to the library function pointer,
and it came up as the use instead of the containing call.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3706
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/26952003