PNaCl: SJLJ EH: Fix incorrect use of setjmp()
While adding eh_throw_tests.cc to tests/toolchain/, I discovered that
test_dtor() would fail with PNaCl's SJLJ exception handling if
optimisation is enabled.  The test does this:

  bool caught = false;
  try {
    func_with_dtor(&caught);
  } catch (MyException &) {
    assert(caught);
  }

PNaClSjLjEH would transform that into something like this:

  bool caught = false;
  if (!setjmp(jmp_buf_for_frame...)) {
    func(&caught);
  } else {
    assert(caught);
  }

But that is incorrect because, if 'caught' is modified by func(),
'caught' will have an undefined value after calling longjmp().  LLVM's
GVN pass was converting 'assert(caught)' to 'assert(false)', because
it doesn't know (and it doesn't need to know) that setjmp() can return
twice.  This broke the original program.

We fix this by moving the setjmp() and func() calls into a separate
helper function.  This means 'caught' won't be local to the function
calling setjmp(), which avoids this quirk in how setjmp()/longjmp()
are specified.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3696
TEST=pnacl-sjlj-eh.ll
     pnacl-sjlj-eh-bug.ll (demonstrates bug at IR level)
     run_eh_throw_tests_sjlj_test (end-to-end test)

Review URL: https://codereview.chromium.org/92543002