Remove some unused PNaCl localmods related to ELF symbol versioning
These were part of PNaCl's old glibc/ELF-based dynamic linking
support, which we are removing.

Normally, symbols with an "@" version will never appear in LLVM IR,
because symbol versions are added by linker when linking an ELF DSO or
executable.  So this localmod is not useful in upstream LLVM.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3520
TEST=trybots

Review URL: https://codereview.chromium.org/120093003