PNaCl: Fix a use-after-free in the ExpandIndirectBr pass
Destroying uses of a function while iterating over the uses leads to a
use-after-free(), which Valgrind detects.  Fix this by making a
temporary copy of the use list.

This bug is probably what was causing the expand-indirectbr.ll test to
crash on Windows (on Cygwin) on the PNaCl toolchain bots.
Specifically, the @unused_blockaddress subtest was crashing on Cygwin.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3695
TEST=expand-indirectbr.ll under Valgrind
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/133843005