PNaCl: Fix ReplacePtrsWithInts' handling of llvm.invariant.end
Before this change, replace-ptrs-with-ints.ll makes 'opt' abort with:

  Attribute after last parameter!
  void (i64, i8*)* @llvm.invariant.end
  Broken module found, compilation aborted!

In LLVM 3.3, llvm-lit doesn't check opt's exit status, and this abort
happens after printing the transformed IR, so this goes unnoticed.  In
LLVM 3.4, llvm-lit is stricter.

This abort happens because the test declares llvm.invariant.end
wrongly.  It actually has 3 args, not 2.  Fixing this means we also
have to fix the pass's handling of llvm.invariant.end.
llvm.invariant.end refers to the result of llvm.invariant.start, so we
have to remove calls to these in the right order.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3757
TEST=tested with LLVM 3.4 merged
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/143803003