Disable more unnecessary PNaCl passes.
To help this, JSBackend.cpp is changed to use DataLayout to get a type
size instead of relying on getPrimitiveSizeInBits, so that it works with
pointer types, and the ExpandI64 pass is changed to not modify the
llvm.lifetime intrinsics.

And this changes some const Type* to plain Type*. Type is immutable, so
many APIs do not support using const with it.