Drop the SimplifyAllocas pass.
The JS backend doesn't need to worry as much about types in the way that the CPP
backend did, so all this pass is doing is saving a little stack space in rare
cases. If these stack savings are important, this should really be handled in
the optimizer proper rather than in a JSBackend-specific pass.