Normalize abbreviation ids in pnacl-bccompress.
Pnacl-bccompress reads a pnacl bitcode file and builds distribution
maps from the records in the file. This includes mapping abbreviations
indices in the records. However, this tool converts all abbreviations
to global, including local abbreviations, which are not unique. This
code fixes the distribution maps so they all refer to the (new)
globalized abbreviations.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3720
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/130913012