Create a new context for main thread instead of using the global context
This is to work around a bug in lib/CodeGen/PseudoSourceValue.cpp which
gets a type from the global context rather than using the context of its
module. It uses a lock to synchronize its own use of the context, but
races with any other code's use.
Rather than plumb the type or context that it needs through all the
necessary layers, we use a new context instead of the global context
for the main module, leaving PseudoSourceValue as the only user of
the global context. This is sufficient for PseudoSourceValue's limited
use of the context, and does not require invasive changes.

We should consider fixing this upstream.

R=jvoung@chromium.org, jfb@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3777

Review URL: https://codereview.chromium.org/156903003