Add threading support for sandboxed SRPC translation
Add a new RPC, StreamInitWithSplit, to initiate the module-splitting
code path, and pass as many descriptors as necessary. To keep the
translator code simple, all the descriptors are passed in one RPC.

Introduce ThreadedStreamingCache to manage the threads' access to the
streaming bitcode object.

Instead of passing a DataStreamer to the bitcode reader and having the
bitcode reader allocate the StreamingMemoryObject, just pass a
StreamableMemoryObject to it directly. This keeps its logic unchanged
and allows the caller to set up the shared streaming

R=jfb@chromium.org, jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3777

Review URL: https://codereview.chromium.org/159753005