Prep for merging 3.4: Strip unrecognised attributes instead of complaining
This is in preparation for the LLVM 3.4 merge.  I found I had to add
"Cold" and "Builtin" to the list of attributes to strip.

When I originally wrote StripAttributes, it seemed like a good idea to
force OK-to-remove attributes to be listed, but now it seems like a
nuisance, because the vast majority of attributes are just hints.
It's a nuisance to have to update StripAttributes' list after a merge.

If any new attributes appear that break programs when naively removed,
it should be OK to rely on testing to discover that.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3757
TEST=tested earlier version with 3.4 merged
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/166373009