Make ReduceMemoryFootprint the default when streaming (w/ LazyBitcode).
This overlaps mostly with LazyBitcode, and LazyBitcode
is flipped on in most cases we care about. Also reducing
memory footprint is a good thing to have by default
(already set by default for pnacl-translate.py, but
not for in-browser translator).

The driver change https://codereview.chromium.org/178043007/
must go in before this, otherwise there will still be
a reference to the flag. See driver CL description for more
notes...

BUG=none
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/177073023