Speed up abbreviation index selection in pnacl-bccompress.
The old implementation would try every abbreviation in a block to
find the best abbreviation to apply. This works well when there
are only a few dozen abbreviations. However, after applying
pnacl-bccompress, the number of abbreviations can exceed 1000 for
a block. Hence, a better solution is needed. This code builds
tries, based on constants that appear in abbreviations, to filter
down what abbreviations can be applied to a bitcode record.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3720
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/183513002